refactor(report): 거대 파일 분리 - ReportApplyPage 섹션 추출 + BottomSheet 가격 로직 hook화 (4/N)
- ReportApplyPage(661줄)를 apply/ 하위로 컴포넌트 분리 (661→194줄)
  - apply/{CallOut,ProgramInfoSection,DocumentSection,PremiumSection,
    ScheduleSection,AdditionalInfoSection,FileUploadButton}.tsx
  - 페이지 본체는 ReportApplyPage.tsx에 유지, 외부 importer(application 페이지)
    import 경로를 apply/*로 갱신
- ReportApplyBottomSheet(764줄): 가격/옵션 파생 계산을
  hooks/useReportApplyPricing로 추출 (764→679줄). 라디오 플랜 선택 로직은
  렌더와 강결합이라 컴포넌트에 유지
