+++ apps/web/src/domain/report/apply/AdditionalInfoSection.tsx
@@ -1,0 +1,50 @@
+'use client';
+
+import React from 'react';
+
+import FilledInput from '@/domain/report/ui/FilledInput';
+import Heading2 from '@/domain/report/ui/heading/Heading2';
+import Label from '@/domain/report/ui/heading/Label';
+import RequiredStar from '@/domain/report/ui/RequiredStar';
+import useReportApplicationStore from '@/store/useReportApplicationStore';
+
+export const AdditionalInfoSection = () => {
+  const { data, setReportApplication } = useReportApplicationStore();
+
+  const onChange = (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
+  ) => {
+    setReportApplication({ [e.target.name]: e.target.value });
+  };
+
+  return (
+    <section className="flex flex-col gap-5">
+      <Heading2>추가 정보</Heading2>
+      <div>
+        <Label htmlFor="wishJob">
+          희망직무
+          <RequiredStar />
+        </Label>
+        <FilledInput
+          name="wishJob"
+          id="wishJob"
+          placeholder="희망하는 직무를 알려주세요"
+          value={data.wishJob || ''}
+          onChange={onChange}
+        />
+      </div>
+      <div>
+        <Label htmlFor="message">서류 작성 고민</Label>
+        <textarea
+          id="message"
+          className="bg-neutral-95 text-xsmall14 w-full resize-none rounded-md p-3"
+          name="message"
+          placeholder="진단에 참고할 수 있도록 서류 작성에 대한 고민을 적어주세요"
+          rows={2}
+          value={data.message || ''}
+          onChange={onChange}
+        />
+      </div>
+    </section>
+  );
+};
