+++ apps/web/src/domain/report/apply/FileUploadButton.tsx
@@ -1,0 +1,56 @@
+'use client';
+
+import React, { useEffect, useRef } from 'react';
+import { IoCloseOutline } from 'react-icons/io5';
+
+export const FileUploadButton = React.memo(function FileUploadButton({
+  file,
+  dispatch,
+}: {
+  file: File | null;
+  dispatch: React.Dispatch<React.SetStateAction<File | null>>;
+}) {
+  const ref = useRef<HTMLInputElement>(null);
+
+  useEffect(() => {
+    if (file && file.size > 50 * 1024 * 1024) {
+      // 파일 사이즈가 50MB 초과일 경우
+      alert('50MB 이하의 파일만 업로드 가능합니다.');
+      dispatch(null);
+    }
+  }, [file, dispatch]);
+
+  return (
+    <>
+      <button
+        className="border-neutral-60 text-xsmall14 text-neutral-40 rounded-md border bg-neutral-100 px-3 py-1.5"
+        onClick={() => {
+          ref.current?.click();
+        }}
+      >
+        {file ? (
+          <div className="flex items-center gap-1">
+            <span>{`${file.name} (${(file.size / 1024).toFixed(1)}KB)`}</span>
+            <IoCloseOutline
+              size={16}
+              color="#7a7d84"
+              onClick={(e) => {
+                e.stopPropagation();
+                dispatch(null);
+              }}
+            />
+          </div>
+        ) : (
+          '파일 업로드'
+        )}
+      </button>
+      <input
+        onChange={(e) => dispatch(e.target.files![0])}
+        className="hidden"
+        ref={ref}
+        type="file"
+        name="file"
+      />
+    </>
+  );
+});
